refactor: update references from EN 1993-1-1:2025 to EN 1993-1-1:2022 in formulas and tests
